Added toggle to Show/Hide code for calendar component - Removed accordian and used toggle for that - Added tabs to display HTML & TS code - Copy code functionality for HTML & TypeScript
