+++ apps/ptg-angular-app/src/app/dashboard/calendarexample/calendarexample.component.scss
@@ -1,0 +1,36 @@
+.parent-block {
+  position: relative;
+}
+
+pre {
+  border: 1px solid #e6e5e5;
+}
+
+.icon-block {
+  cursor: pointer;
+  align-items: center;
+  background: #f2f2f2;
+  border-radius: 50%;
+  display: flex;
+  height: 50px;
+  justify-content: center;
+  position: absolute;
+  right: 80px;
+  top: 0;
+  width: 50px;
+
+  &:hover {
+    background: lightgray;
+  }
+}
+
+::ng-deep .tab-content {
+  position: relative;
+
+  i.fa-copy {
+    cursor: pointer;
+    position: absolute;
+    right: 90px;
+    top: 40px;
+  }
+}
