--- apps/ptg-angular-app/src/app/dashboard/calendarexample/calendarexample1/calendarexample1.component.scss
@@ -1,21 +1,0 @@
-.panel-open ::ng-deep .collapse.in {
-  position: relative;
-}
-
-.panel-open ::ng-deep .collapse.in i {
-  cursor: pointer;
-  padding: 15px;
-  position: absolute;
-  right: 20px;
-  top: 15px;
-
-  &:hover {
-    background: lightgray;
-    border-radius: 50%;
-    padding: 15px;
-  }
-}
-
-.pb-2 ::ng-deep .btn.btn-link {
-  text-decoration: none;
-}
