+++ apps/ptg-angular-app/src/app/dashboard/calendarexample/calendarexample1/calendarexample1.component.ts
@@ -15,9 +15,7 @@ import { ChangeDetectorRef, Component, OnInit, AfterViewInit } from '@angular/co
 import { FormBuilder, FormGroup, Validators } from '@angular/forms';
 import { calendarValidator } from '@ptg-angular-app/common/utils/validators';
 import { resources } from "../../../../resource/resource";
-import { MatSnackBar } from '@angular/material/snack-bar';
 
-import { ACCORDIAN_DATA } from '@ptg-angular-app/mock/mocks';
 @Component({
   selector: 'ptg-ui-calendarexample1',
   templateUrl: './calendarexample1.component.html',
@@ -30,28 +28,6 @@ export class Calendarexample1Component implements OnInit, AfterViewInit {
   endMinDate = new Date();
   resources=resources
   submitted = false
-  data = ACCORDIAN_DATA;
-  htmlCode = `
-    <ptg-ui-calendar [minDate]="minDate" [maxDate]="endDate" placeholder='MM-DD-YYYY' className='' id=''
-                     [disabled]='false' [readOnly]="false" (calendarValueChange)="onDateChange($event)">
-    </ptg-ui-calendar>`;
-
-  tsCode =
-    `
-    import { Component } from '@angular/core';
-
-    @Component({
-      selector: 'demo-calendar-component',
-      templateUrl: './demo-calendar.component.html'
-    })
-    export class DemoCalendarComponent {
-      minDate = new Date();
-      maxDate = new Date();
-
-      constructor() {
-        this.maxDate.setDate(this.minDate.getDate() + 30);
-      }
-    }`
 
   get f() {
     return this.calendarForm.controls;
@@ -61,7 +37,7 @@ export class Calendarexample1Component implements OnInit, AfterViewInit {
     return this.calendarForm.get('endDate');
   }
 
-  constructor(private formBuilder: FormBuilder, private cdr: ChangeDetectorRef, private _snackbar: MatSnackBar) { }
+  constructor(private formBuilder: FormBuilder, private cdr: ChangeDetectorRef) { }
 
   ngOnInit(): void {
     this.calendarForm = this.formBuilder.group({
@@ -79,13 +55,6 @@ export class Calendarexample1Component implements OnInit, AfterViewInit {
 
   ngAfterViewInit(){
     this.cdr.detectChanges();
-  }
-
-  openSnackBar() {
-    this._snackbar.open('Code Copied', '', {
-      verticalPosition: 'bottom',
-      horizontalPosition: 'center'
-    })
   }
 
   onStartDateChange(event:any){
